refactor: move stopContainer to containerutil
